Wrap self in sudo if not root
This is for systems which make it a pain to run sudo with stuff not in
the root user's default path.

For example on Solus:

~~~
$ inin up
Program terminated.
You have to be root for this operation.
Please use 'eopkg help' for general help.

$ sudo inin up
Password:
sudo: inin: command not found
~~~

I didn't want to wrap each line in sudo, partially for systems which
don't have sudo, but also it's just messy and means for long running
commands that you might have to re-authentice multiple times.

Alpine docker images don't ship with sudo by default and you run as root
already, causing double trouble. (Alpine isn't supported yet, but will
be.)